+++ lib/worker/worker_middleware.dart
@@ -4,7 +4,13 @@ import 'package:meowchannel/extensions/channel.dart';
 import 'package:meowchannel/meowchannel.dart';
 import 'package:meowchannel/worker/watcher.dart';
 import 'package:meowchannel/worker/worker_context.dart';
- 
+
+class SetWatchersAction<S> {
+  List<Watcher<dynamic, S>> watchers;
+
+  SetWatchersAction(this.watchers);
+}
+
 Middleware workerMiddleware<S>(
   List<Watcher<dynamic, S>> watchers
 ) => (
@@ -12,22 +18,20 @@ Middleware workerMiddleware<S>(
   Function() getState,
   Dispatcher next
 ) {
-  final context = WorkerContext<S>(
-    dispatcher: dispatcher,
-    state: getState
-  );
+  final context = WorkerContext<S>(dispatcher, getState);
   final channel = StateChannel<dynamic>();
-  
+
   watchers.forEach((Watcher<dynamic, S> watcher) {
     watcher.watch(channel.asStream(), context);
   });
 
   return (dynamic action) async {
-    if(action is MeowChannelClose)
+    if(action is MeowChannelClose) {
       channel.close();
-    else
+    } else {
       channel.send(action);
-      
+    }
+
     await next(action);
   };
 };