+++ test/1_basic_store/root_state.g.dart
@@ -22,7 +22,7 @@ abstract class _$RootState {
   }
 
   String toString() {
-    return 'RootState <\'value\': ${this.value},>';
+    return 'RootState(value=${this.value},)';
   }
 
   RootState copyWith({int value}) {